requirements.txt
streamlit==1.28.0
numpy==1.23.5
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.1.78
tensorflow==2.13.0
matplotlib==3.7.2
Pillow==10.0.1